// src/redux/products/products.types.ts
enum ProductsActionTypes {
  FETCH_ALL_PRODUCTS_REQUESTED = 'FETCH_ALL_PRODUCTS_REQUESTED',
  FETCH_ALL_PRODUCTS_SUCCESS = 'FETCH_ALL_PRODUCTS_SUCCESS',
  FETCH_ALL_PRODUCTS_FAILED = 'FETCH_ALL_PRODUCTS_FAILED',

  FETCH_PRODUCT_REQUESTED = 'FETCH_PRODUCT_REQUESTED',
  FETCH_PRODUCT_FOR_EDIT_REQUESTED = 'FETCH_PRODUCT_FOR_EDIT_REQUESTED',
  FETCH_PRODUCT_SUCCESS = 'FETCH_PRODUCT_SUCCESS',
  FETCH_PRODUCT_FAILED = 'FETCH_PRODUCT_FAILED',

  FETCH_SINGLE_COLLECTION_PRODUCTS_REQUESTED = 'FETCH_SINGLE_COLLECTION_PRODUCTS_REQUESTED',
  FETCH_SINGLE_COLLECTION_PRODUCTS_SUCCESS = 'FETCH_SINGLE_COLLECTION_PRODUCTS_SUCCESS',
  FETCH_SINGLE_COLLECTION_PRODUCTS_FAILED = 'FETCH_SINGLE_COLLECTION_PRODUCTS_FAILED',

  CREATE_PRODUCT_REQUESTED = 'CREATE_PRODUCT_REQUESTED',
  CREATE_PRODUCT_SUCCESS = 'CREATE_PRODUCT_SUCCESS',
  CREATE_PRODUCT_FAILED = 'CREATE_PRODUCT_FAILED',

  EDIT_PRODUCT_REQUESTED = 'EDIT_PRODUCT_REQUESTED',
  EDIT_PRODUCT_SUCCESS = 'EDIT_PRODUCT_SUCCESS',
  EDIT_PRODUCT_FAILED = 'EDIT_PRODUCT_FAILED', 

  DELETE_PRODUCT_REQUESTED = 'DELETE_PRODUCT_REQUESTED',
  DELETE_PRODUCT_SUCCESS = 'DELETE_PRODUCT_SUCCESS',
  DELETE_PRODUCT_FAILED = 'DELETE_PRODUCT_FAILED',

  ADD_COLOR_TO_PRODUCT_REQUESTED = 'ADD_COLOR_DATA_TO_PRODUCT_REQUESTED',
  ADD_COLOR_TO_PRODUCT_SUCCESS = 'ADD_COLOR_DATA_TO_PRODUCT_SUCCESS',
  ADD_COLOR_TO_PRODUCT_FAILED = 'ADD_COLOR_DATA_TO_PRODUCT_FAILED',

  EDIT_PRODUCT_COLOR_DATA_REQUESTED = 'EDIT_PRODUCT_COLOR_DATA_REQUESTED',
  EDIT_PRODUCT_COLOR_DATA_SUCCESS = 'EDIT_PRODUCT_COLOR_DATA_SUCCESS',
  EDIT_PRODUCT_COLOR_DATA_FAILED = 'EDIT_PRODUCT_COLOR_DATA_FAILED',

  REMOVE_ONE_IMAGE_FROM_COLOR_IMAGES_REQUESTED = 'REMOVE_ONE_IMAGE_FROM_COLOR_IMAGES_REQUESTED',
  REMOVE_ONE_IMAGE_FROM_COLOR_IMAGES_SUCCESS = 'REMOVE_ONE_IMAGE_FROM_COLOR_IMAGES_SUCCESS',
  REMOVE_ONE_IMAGE_FROM_COLOR_IMAGES_FAILED = 'REMOVE_ONE_IMAGE_FROM_COLOR_IMAGES_FAILED',

  DELETE_COLOR_FROM_PRODUCT_REQUESTED = 'DELETE_COLOR_FROM_PRODUCT_REQUESTED',
  DELETE_COLOR_FROM_PRODUCT_SUCCESS = 'DELETE_COLOR_FROM_PRODUCT_SUCCESS',
  DELETE_COLOR_FROM_PRODUCT_FAILED = 'DELETE_COLOR_FROM_PRODUCT_FAILED'
}

export default ProductsActionTypes